TodoApp/backend/pyproject.toml
[project]
name = "todo-api"
version = "0.1.0"
description = "A modern RESTful API built with FastAPI for managing todo items"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=1.4,<2.0",
    "alembic>=1.12.0",
    "pydantic>=2.4.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "jinja2>=3.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.6",
    "bcrypt>=4.0.1",
    "passlib[bcrypt]>=1.7.4"
]
requires-python = ">=3.8"